% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedr.R
\name{add_background}
\alias{add_background}
\title{add_background}
\usage{
add_background(gene_array)
}
\arguments{
\item{gene_array}{}
}
\value{
array
}
\description{
Upload a background gene list. The background will define the scope of all possible genes. All uploaded gene sets and gene sets in gene set libraries will first be filtered by the background before calculating significance of overlap.
}
\examples{
libraries <- speedr::add_background(c("SOX2", "NANOG", "RUNX2", "GATA4"))

}
