% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addResults.R
\name{addResults}
\alias{addResults}
\title{Add Results to Master Data}
\usage{
addResults(file)
}
\arguments{
\item{file}{The name of the individual race results .csv file}
}
\description{
This function adds race results to the master results table.
}
\examples{
addResults("Data/Results/mRegionals15.csv")
}
\keyword{country}
\keyword{cross}
\keyword{results,}
