% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSpeedRatings.R
\name{addSpeedRatings}
\alias{addSpeedRatings}
\title{Save Speed Ratings}
\usage{
addSpeedRatings(speedRankings, location = "Data/Speed Ratings/")
}
\arguments{
\item{speedRankings}{A list of speed rating data frames}

\item{location}{The folder destination you want to save your output}
}
\description{
This function saves speed rating output to a designated data folder.
}
\examples{
addSpeedRatings(speedRankings)
}
\keyword{cross}
\keyword{crountry,}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
