% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap race results to generate reference estimate}
\usage{
bootstrap(allSpeedRatings, year)
}
\arguments{
\item{allSpeedRatings}{The master speed rating data frame}

\item{year}{The year of the current race}
}
\value{
The data frame of updated reference speed ratings with an uncertainty measure taking into account weighting scheme
}
\description{
This function inputs a master file of speed ratings and references and returns a list of updated reference speed ratings for the runners contained in the master list.
}
\examples{
bootstrap(allSpeedRatings, 2018)
}
\keyword{country,}
\keyword{cross}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
