% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseCorrectionDiagPlot.R
\name{courseCorrectionDiagPlot}
\alias{courseCorrectionDiagPlot}
\title{Visualize Quality of Estimates}
\usage{
courseCorrectionDiagPlot(results, references, alpha = 4.4, beta = 2355,
  gamma)
}
\arguments{
\item{results}{The results for the race you are speed rating}

\item{alpha}{The number of seconds one speed rating point is equal to}

\item{beta}{The number of seconds marking the zero point of the scale}

\item{gamma}{The course correction parameter}

\item{reference}{The data frame of expected performances}
}
\value{
A plot of expected speed rating against time with a line visualizing estimated speed ratings
}
\description{
This function estimates and returns the course correction parameter, gamma, for courses that are not the base course.
}
\examples{
getCourseCorrection(results = geneseo16, alpha = 4.4, beta = 2355, lower_thresh = 0.1, upper_thresh = 0.95, baseID = "mGeneseo15", baseIntercept = 1527.197)
}
\keyword{country,}
\keyword{cross}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
