% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedr.R
\name{enrich}
\alias{enrich}
\title{enrichr}
\usage{
enrich(library, gene_array, background = NA)
}
\arguments{
\item{library}{}

\item{gene_array}{}
}
\value{
array
}
\description{
Run enrichment analysis for a given gene set and gene set library.
}
\examples{
# enrichment analysis without background correction
library = "GO_Biological_Process_2021"
genes = c('PTPN18','EGF','HSP90AA1','GAB1','NRG1','MATK','PTPN12','NRG2','PTK6','PRKCA','ERBIN','EREG','BTC','NRG4','PIK3R1','PIK3CA','CDC37','GRB2','STUB1','HBEGF','GRB7')
result <- speedr::enrich(library, genes)

# enrichment analysis with background correction
library = "GO_Biological_Process_2021"
background = c('TGFB1', 'EOMES', 'TGFB2', 'DDX17', 'NOG', 'FOXF2', 'WNT5A', 'HGF', 'HMGA2', 'HNRNPAB', 'PTPN18','EGF','HSP90AA1','GAB1','NRG1','MATK','PTPN12','NRG2','PTK6','PRKCA','ERBIN','EREG','BTC','NRG4','PIK3R1','PIK3CA','CDC37','GRB2','STUB1','HBEGF','GRB7')
genes = c('TGFB1', 'EOMES', 'TGFB2', 'DDX17', 'NOG', 'FOXF2', 'WNT5A', 'HGF', 'HMGA2', 'HNRNPAB', 'PTPN18','EGF','HSP90AA1','GAB1','NRG1','MATK','PTPN12','NRG2','PTK6','PRKCA','ERBIN','EREG','BTC','NRG4','PIK3R1','PIK3CA','CDC37','GRB2','STUB1','HBEGF','GRB7')
result <- speedr::enrich(library, genes, background)

}
