% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSpeedRankings.R
\name{estimateSpeedRankings}
\alias{estimateSpeedRankings}
\title{Estimate Speed Ratings}
\usage{
estimateSpeedRankings(race, results, courseCorrections, alpha = 4.4,
  beta = 2355)
}
\arguments{
\item{race}{The race you would like to calculate speed ratings for}

\item{results}{The master results data frame}

\item{courseCorrections}{The master course corrections data frame}

\item{alpha}{The number of seconds equal to one speed point}

\item{beta}{The centering parameter}
}
\description{
This function estimates speed ratings for a course with a course correction 
and returns an easy to interpret data frame of results with the estimated 
speed ratings.
}
\examples{
estimateSpeedRankings("9349", results, courseCorrections)
}
\keyword{cross}
\keyword{crountry,}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
