% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseo15_metrics.R
\docType{data}
\name{geneseo15_metrics}
\alias{geneseo15_metrics}
\title{Geneseo '15 Race Metrics}
\format{An object of class \code{"data.frame"}.}
\usage{
data(geneseo15_metrics)
}
\description{
Race measures for mGeneseo15 used as baseline comparison when determining
course corrections.
}
\keyword{data}
