% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCourseCorrection_avgDiff.R
\name{getCourseCorrection_avgDiff}
\alias{getCourseCorrection_avgDiff}
\title{Get Course Correction Parameter Using Average Difference from Expectation}
\usage{
getCourseCorrection_avgDiff(results, referenceRunners, alpha = 4.4,
  beta = 2355, lower_thresh = 0.05, upper_thresh = 0.95,
  baseID = "mGeneseo15")
}
\arguments{
\item{race}{The numeric ID of the race you want to estimate the correction for}
}
\value{
gammaFit The best fit estimate for the course correction parameter, gamma.
}
\description{
This function estimates and returns the course correction parameter, gamma, for courses that are not the base course.
}
\examples{
getCourseCorrection_avgDiff(race = "mWilliams15", results, referenceRunners, guess, baseID = "mGeneseo15")
}
\keyword{country,}
\keyword{cross}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
