% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpdatedRefSR.R
\name{getUpdatedRefSR}
\alias{getUpdatedRefSR}
\title{Get Updating Reference Speed Rating}
\usage{
getUpdatedRefSR(allSpeedRatings, races)
}
\arguments{
\item{allSpeedRatings}{The master speed rating data frame}

\item{races}{The list of races in chronological order}
}
\value{
updatedReferences The data frame of updated reference speed ratings
}
\description{
This function inputs a master file of speed ratings for a year and a list of races and returns a list of updated reference speed ratings for the runners contained in the master list.
}
\examples{
getUpdatedRefSR(allSpeedRatings, c("mOswego17", "mWilliams17", "mGeneseo17"))
}
\keyword{country,}
\keyword{cross}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
