% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedr.R
\name{import_gmt}
\alias{import_gmt}
\title{import_gmt}
\usage{
import_gmt(gmt_name, gmt)
}
\arguments{
\item{gmt}{}
}
\value{
status
}
\description{
Upload a GMT file.
}
\examples{

# download GMT file from Enrichr database
url = "https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=KEGG_2021_Human"
download.file(url, "KEGG_2021_Human.gmt")
gmt <- speedr::read_gmt("KEGG_2021_Human.gmt")

# push GMT file to API
res <- speedr::import_gmt("KEGG_2021_Human", gmt)

}
