% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedr.R
\name{list_libraries}
\alias{list_libraries}
\title{list_libraries}
\usage{
list_libraries()
}
\value{
array
}
\description{
List libraries hosted by Speedrichr instance.
}
\examples{
libraries <- speedr::list_libraries()

}
