% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_course_correction.R
\docType{data}
\name{m_gam}
\alias{m_gam}
\title{Generalied Additive Model for Predicting Course Corrections}
\format{An object of class \code{"gam", "glm", "lm"}.}
\usage{
data(gam_course_correction)
}
\description{
Model trained on 74 races with human coded course corrections. Model predicts
course correction from the difference between the baseline race and the new
race at the 10th, 20th, 30th, 40th, and 50th percentiles, as well as the new
race's skewness, kurtosis, and standard deviation.
}
\keyword{model}
