% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCourseCorrection.R
\name{predictCourseCorrection}
\alias{predictCourseCorrection}
\title{Get Course Correction Parameter}
\usage{
predictCourseCorrection(results, race_dist = "8k", model_type = "lm")
}
\arguments{
\item{results}{The results for the race you are speed rating}

\item{race_dist}{The distance of the race being analyzed}

\item{model_type}{Identifies whether to use linear or additive model for prediction}
}
\value{
gamma The predicted course correction
}
\description{
This function estimates and returns the course correction parameter, gamma, for courses that are not the base course.
}
\examples{
predictCourseCorrection(results = geneseo16, race_dist = "8k", model_type = "lm")
}
\keyword{country,}
\keyword{cross}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
