% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedr.R
\name{read_gmt}
\alias{read_gmt}
\title{read_gmt}
\usage{
read_gmt(filename)
}
\arguments{
\item{filename}{}
}
\value{
list
}
\description{
Read GMT file to list for later import to API.
}
\examples{

# download GMT file from Enrichr database
url = "https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=KEGG_2021_Human"
download.file(url, "KEGG_2021_Human.gmt")

gmt <- speedr::read_gmt("KEGG_2021_Human.gmt")

}
