% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readResults.R
\name{readResults}
\alias{readResults}
\title{Read Race Results}
\usage{
readResults(file)
}
\arguments{
\item{file}{The location of individual race results in .csv format}
}
\value{
A tibble of race results with incomplete rows and extra columns removed.
}
\description{
Load and clean race results data in .csv format.
}
\examples{
readResults("Data/Results/mRowanV15.csv")
}
