% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR.R
\name{SR}
\alias{SR}
\title{Baseline Speed Rating Estimation}
\usage{
SR(x, alpha, beta)
}
\arguments{
\item{x}{Time in seconds for runners as a numeric vector}

\item{alpha}{The number of seconds equal to one speed point}

\item{beta}{The centering parameter}
}
\description{
This function estimates speed ratings for a baseline course.
}
\examples{
SR(1573.6, 4.4, 2355)
}
\keyword{cross}
\keyword{crountry,}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
