% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_CourseCorrection.R
\name{SR_CourseCorrection}
\alias{SR_CourseCorrection}
\title{Course Corrected Speed Rating Estimation}
\usage{
SR_CourseCorrection(x, alpha = 4.4, beta = 2355, gamma = 0)
}
\arguments{
\item{x}{Time in seconds for runners as a numeric vector}

\item{alpha}{The number of seconds equal to one speed point}

\item{beta}{The centering parameter}

\item{gamma}{The course correction parameter indicating estimated average difference between a course and the baseline}
}
\description{
This function estimates speed ratings for a course with a course correction.
}
\examples{
SR_CourseCorrection(1595.2, 4.4, 2355, -46)
}
\keyword{cross}
\keyword{crountry,}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
