% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCourseCorrection.R
\name{getCourseCorrection}
\alias{getCourseCorrection}
\title{Get Course Correction Parameter}
\usage{
getCourseCorrection(race, results, referenceRunners, alpha = 4.4,
  beta = 2355, baseID = "8306")
}
\arguments{
\item{race}{The numeric ID of the race you want to estimate the correction for}
}
\value{
gammaFit The best fit estimate for the course correction parameter, gamma.
}
\description{
This function estimates and returns the course correction parameter, gamma, for courses that are not the base course.
}
\examples{
getCourseCorrection(9198, 8)
}
\keyword{country,}
\keyword{cross}
\keyword{handicapping}
\keyword{rating,}
\keyword{speed}
